Renamed package @blocks/blocks to @blocks/react
