+++ packages/blocks/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@blocks/blocks",
+  "name": "@blocks/react",
   "version": "0.0.41",
   "description": "The default, Theme UI-based for the Blocks editor",
   "repository": "blocks/blocks",
@@ -15,7 +15,6 @@
     "@babel/preset-react": "7.7.4"
   },
   "dependencies": {
-    "@blocks/blocks": "^0.0.41",
     "property-controls": "^0.0.41"
   }
 }
