+++ packages/blocks/readme.md
@@ -1,4 +1,4 @@
-# @blocks/blocks
+# @blocks/react
 
 The default, Theme UI-based for the Blocks editor.
