+++ src/components/block.js
@@ -4,7 +4,7 @@ import { Link } from 'gatsby'
 import { jsx, Styled } from 'theme-ui'
 import * as components from '@theme-ui/components'
 import { Global } from '@emotion/core'
-import * as blocks from '@blocks/blocks/src'
+import * as blocks from '@blocks/react/src'
 import InlineRender from 'blocks-ui/dist/inline-render'
 import * as controls from 'property-controls'
 import SEO from './seo'
