+++ src/pages/getting-started.mdx
@@ -34,7 +34,7 @@ import Editor from 'blocks-ui'
 
 const JSX = `
 import React from 'react'
-import { HeaderBasic } from '@blocks/blocks'
+import { HeaderBasic } from '@blocks/react'
 
 export default () => (
   <Blocks.Root>
